requirements.txt
SQLAlchemy~=2.0.23
fastapi~=0.105.0
uvicorn~=0.25.0
pytest~=7.4.4
alembic~=1.13.1
pydantic~=2.5.3
python-dotenv~=1.0.0
async-asgi-testclient~=1.4.11
pydantic-settings~=2.1.0
asyncpg~=0.29.0
pyyaml~=6.0.1